Add script to download Errantry Calc & Tech List code
The additional data-points for techniques are scraped from the JavaScript
source code of LegendCup.com's Errantry Calculator & Tech List page.
This code rarely changes, but gets redownloaded every time the
scrape-techniques.ps1 script or various entity classes and scripts
change.

Data for other tables will also need to be scraped from this code as
well, so use a separate script to download it, that way it only has to
be downloaded whenever the download script or file utilities script
changes.