Start IDs at 1 for most tables
The `sqlite3_last_insert_rowid` API returns 0 if nothing has ever been
inserted into the table, making 0 a 0 a de-facto "invalid" rowid.
Source: https://stackoverflow.com/questions/8966667/valid-range-of-sqlite-rowid

It is better to use 1 as the starting row ID. Will do the same with the
Breeds table in a later commit as it will require more work.